sistema-educacional/requirements.txt
websocket-client==1.8.0
websocket-server==0.6.4
